+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 -   Fixed issue with addChild if the child has already been added
 -   Added position type for relative/absolute positioning
 -   Added support for percentage based x/y
+-   Updated documentation and added logger options to the main initialize fuunction
 
 # 0.3.0